Add basic rendering function and server communication
Socket.io is properly integrated now, and there is now a method res.screen that is created by middleware that will render your view and then put it into the screen template. You can still, however, use res.render or res.end if so desired.

Keystone is now mounted using keystone.mount which allows us to have finer control of the routing and middleware. It is only under /as/, so we can render pretty views and such there, but everywhere else it is bare bones for maximum control.

Namespace for all public assets should be public/namespace