Add dashboard to /as/
The dashboard is a temporary solution to Keystone's shortcomings. Because of this, not a heck of a lot effort will be put into making it clean. It exists solely so AtriumScreen can be tested properly while Keystone is developed.

It currently allows for new pages to be defined using atriumscreen.navDash - an array of page names
To define new routes, simply use atriumscreen.on('keystone routes'). Routes here are relative to /as as well
Use res.admin(view, locals) in this to render to the admin UI
You can fill res.locals.scripts and res.local.styles with files you need to link to.

It ain't clean, but it works.