Allow custom scripts and styles in layout controllers
Is simple, just append elements to the res.locals.styles and res.locals.scripts arrays. Please use .push and don't just set them so as to not interfere with middleware.
In the future, there is a plan to stop the usage of two of the same script. Using Bower to do this programatically is probably the best way forward.