Add smooth fade for refresh, allow better scripts
Scripts can now be inline by pushing {src: 'javascript here', inline: true} to a script array.
res.locals.headScripts has been added for scripts that need to be in the <head> element (eg. scripts from Typekit)
On refreshing and loading, a smooth animation plays. There is currently a white flicker as the page actually refreshes, but this will be fixed with the use of iframes.