Put screen functionality in iframe
This allows for the smoothest of refreshes/transitions. What is left to do is allow a greater manipulation of these transitions.

Eventually we will need to determine time sync with server and client. This information (along with other useful persistent information), can be saved in the parent window so it does not have to be recalculated for every transition. For time sensitive applications, this is imperative.