Add continuous refresh when server is unreachable
The master client template now has a tiny bit of code in it. When the client refreshes, it will notify the master, and if in 10 seconds the master has not received a notification that a refresh has been completed from the new client, it will refresh the iframe every 5 seconds.
And this all happens under the as-blockade so is quite seamless even if the server crashes and restarts.